+++ wp-content/themes/the-realm-malta/functions.php
@@ -3,6 +3,7 @@
 require_once('classes/trm-core.php');
 require_once('classes/trm-marketing.php');
 require_once('classes/wc-hooks.php');
+require_once('classes/header-hooks.php');
 require_once('classes/metabox-hooks.php');
 require_once('classes/acf-hooks.php');
 require_once('classes/ajax-hooks.php');
@@ -11,6 +12,7 @@ require_once('classes/profit-analytics/trm-profit-analytics.php');
 
 new TRM_Marketing_Handler();
 new TRM_WC_Hooks();
+new TRM_Header_Hooks();
 new TRM_MB_Hooks();
 new TRM_ACF_Hooks();
 new TRM_AJAX_Hooks();